Moved folder /animation to /scene/animation
